+++ apps/chat/src/utils/attachment-canvas.ts
@@ -1,77 +1,10 @@
 import type {
   ImageCanvasContent,
   PlainTextCanvasContent,
-  UnsupportedCanvasContent,
 } from '@epam/ai-dial-attachment-canvas';
 import { AttachmentContentType } from '@epam/ai-dial-attachment-canvas';
 import type { Attachment, DisplayAttachment } from '@epam/ai-dial-chat-shared';
 import { resolveDialFileDownloadUrl } from './icon-path';
-
-const TEXT_EXTENSIONS = new Set([
-  'txt',
-  'md',
-  'markdown',
-  'json',
-  'jsonl',
-  'ndjson',
-  'xml',
-  'csv',
-  'tsv',
-  'yaml',
-  'yml',
-  'toml',
-  'ini',
-  'conf',
-  'cfg',
-  'html',
-  'htm',
-  'css',
-  'scss',
-  'sass',
-  'less',
-  'js',
-  'jsx',
-  'ts',
-  'tsx',
-  'mjs',
-  'cjs',
-  'mts',
-  'cts',
-  'py',
-  'rb',
-  'go',
-  'rs',
-  'java',
-  'kt',
-  'swift',
-  'c',
-  'h',
-  'cpp',
-  'cs',
-  'sh',
-  'bash',
-  'zsh',
-  'fish',
-  'ps1',
-  'log',
-  'env',
-  'gitignore',
-  'dockerfile',
-  'makefile',
-  'sql',
-]);
-
-/** Returns true if the file name has an extension that is known to be text-previewable. */
-export const isTextPreviewable = (name: string): boolean => {
-  const dot = name.lastIndexOf('.');
-  if (dot === -1) return false;
-  return TEXT_EXTENSIONS.has(name.slice(dot + 1).toLowerCase());
-};
-
-/** Creates an unsupported content payload. */
-export const createUnsupportedCanvasContent = (): UnsupportedCanvasContent => ({
-  type: AttachmentContentType.Unsupported,
-});
 
 /** Resolves an image canvas content payload from a DisplayAttachment, or `null` if unavailable. */
 export const resolveImageCanvasContent = async (
