+++ libs/attachment-canvas/src/constants/file.ts
@@ -1,0 +1,54 @@
+/** File extensions that are known to be previewable as plain text. */
+export const TEXT_EXTENSIONS = new Set([
+  'txt',
+  'md',
+  'markdown',
+  'json',
+  'jsonl',
+  'ndjson',
+  'xml',
+  'csv',
+  'tsv',
+  'yaml',
+  'yml',
+  'toml',
+  'ini',
+  'conf',
+  'cfg',
+  'html',
+  'htm',
+  'css',
+  'scss',
+  'sass',
+  'less',
+  'js',
+  'jsx',
+  'ts',
+  'tsx',
+  'mjs',
+  'cjs',
+  'mts',
+  'cts',
+  'py',
+  'rb',
+  'go',
+  'rs',
+  'java',
+  'kt',
+  'swift',
+  'c',
+  'h',
+  'cpp',
+  'cs',
+  'sh',
+  'bash',
+  'zsh',
+  'fish',
+  'ps1',
+  'log',
+  'env',
+  'gitignore',
+  'dockerfile',
+  'makefile',
+  'sql',
+]);
